Avoid precision issues with casting `RAND_MAX` to `float`

+++ tests/matrix-multiply.c
@@ -4,7 +4,7 @@
 
 float randf(float low, float high) {
     long int r = random();
-    return low + (float)r * ((high - low) / RAND_MAX);
+    return low + (float)((double)r * ((high - low) / (double)RAND_MAX));
 }
 
 void make_random_matrix(float* mat, size_t rows, size_t columns) {
